feat: auth and roles

+++ src/controller/user.py
@@ -1,39 +1,52 @@
 from flask import Blueprint, request
 from src.hello import User, db
-from http import HTTPStatus
+from src.utils import required_role
+from http import HTTPStatus, HTTPMethod
+from flask_jwt_extended import jwt_required, get_jwt_identity
 
 app = Blueprint('user', __name__, url_prefix='/users')
 
-
 def _create_user():
     data = request.json
-    user = User(username=data["username"], password=data["password"])
+    user = User(username=data["username"], password=data["password"], role_id=data["role_id"])
     db.session.add(user)
     db.session.commit()
+
 
 def _list_users():
     query = db.select(User)
     users = db.session.execute(query).scalars().all()
     return [
-        {"id": user.id, "username": user.username} for user in users
+        {
+         "id": user.id, 
+         "username": user.username,
+         "role": {
+             "id": user.role.id,
+             "name": user.role.name
+             }
+         } for user in users
     ]
 
 
-@app.route('/', methods=["GET", "POST"])
-def handler_user():
+@app.route('/', methods=[HTTPMethod.GET, HTTPMethod.POST])
+@jwt_required()
+@required_role("admin")
+def handler_user():   
     if request.method == "POST":
         _create_user()
         return {"message": "User created!"}, HTTPStatus.CREATED
-    else:
+    elif request.method == "GET":
         return {"users":  _list_users()}, HTTPStatus.OK
+    else:
+        return {"message": "Method not allowed"}, HTTPStatus.METHOD_NOT_ALLOWED
         
 
-@app.route('/<int:user_id>', methods=["GET", "PATCH", "DELETE"])
+@app.route('/<int:user_id>', methods=[HTTPMethod.GET, HTTPMethod.PATCH, HTTPMethod.DELETE])
 def handler_user_id(user_id):
-    if request.method == "GET":
+    if request.method == HTTPMethod.GET:
         user = db.get_or_404(User, user_id)
         return {"id": user.id, "username": user.username}, HTTPStatus.OK
-    elif request.method == "PATCH":
+    elif request.method == HTTPMethod.PATCH:
         data = request.json
 
         if "username" in data:
@@ -46,8 +59,10 @@ def handler_user_id(user_id):
         db.session.commit()
 
         return {"id": user.id, "username": user.username}, HTTPStatus.OK
-    else:
+    elif request.method == HTTPMethod.DELETE:
         user = db.get_or_404(User, user_id)
         db.session.delete(user)
         db.session.commit()
         return {"message": "User deleted!"}, HTTPStatus.OK
+    else:
+        return {"message": "Method not allowed"}, HTTPStatus.METHOD_NOT_ALLOWED
